api/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.19 # Ensure compatible version